moving /register to /users for more rest-like api
