--- v0/routes/register/routes.js
@@ -1,6 +1,0 @@
-const { json } = require('../../middleware')
-const registerUser = require('./registerUser')
-
-module.exports = function(app) {
-	app.post('/', json(), registerUser)
-}
