+++ v0/routes/users/routes.js
@@ -1,0 +1,6 @@
+const { json } = require('../../middleware')
+const createUser = require('./createUser')
+
+module.exports = function(app) {
+	app.post('/', json(), createUser)
+}
